Change type of adios2_local_value_dim to size_t

+++ bindings/C/adios2/c/adios2_c_types.h
@@ -11,9 +11,8 @@
 #ifndef ADIOS2_BINDINGS_C_ADIOS2_C_TYPES_H_
 #define ADIOS2_BINDINGS_C_ADIOS2_C_TYPES_H_
 
-#include <limits.h> // ULLONG_MAX
 #include <stddef.h> // size_t
-#include <stdint.h> // uint64_t
+#include <stdint.h> // SIZE_MAX
 
 #include "adios2/common/ADIOSConfig.h"
 
@@ -141,7 +140,7 @@ typedef enum
 
 static const size_t adios2_string_array_element_max_size = 4096;
 
-static const uint64_t adios2_local_value_dim = ULLONG_MAX - 2;
+static const size_t adios2_local_value_dim = SIZE_MAX - 2;
 
 #ifdef __cplusplus
 } // end extern C
